RAG-EKB/requirements.txt
# Web框架和服务器
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.3

# 数据库相关
sqlalchemy==2.0.28
pymysql==1.1.0
cryptography==42.0.5

# 向量模型和机器学习
sentence-transformers==4.1.0
faiss-cpu==1.10.0
numpy>=1.25.0,<3.0
transformers==4.46.2
tokenizers==0.20.3
safetensors==0.5.3

# 文档处理
python-multipart==0.0.20
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
beautifulsoup4==4.13.4
markdown==3.8

# API集成
openai==1.75.0

# 日志和工具
python-json-logger==2.0.7
python-dotenv==1.0.0
